Moved endpoint to /listAllEvents to be more descriptive.  #1
