+++ server/middleware/events/listAllEvents.js
